Remove old components and cleanup structure
