+++ package.json
@@ -25,7 +25,6 @@
     "ethers": "^5.0.14",
     "history": "^5.0.0",
     "node-sass": "^4.14.1",
-    "numbro": "^2.3.1",
     "prettier": "^2.1.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
