--- src/App.scss
@@ -1,186 +1,0 @@
-@import "./styles/theme";
-
-* {
-  border-color: #aaa;
-}
-
-button,
-input,
-select {
-  border-radius: 10px;
-  height: fit-content;
-  min-height: 2em;
-  border-width: thin;
-  background-color: #fff;
-}
-
-.token-selector {
-  display: flex;
-  justify-content: space-around;
-  align-items: flex-end;
-  border: 1px solid #aaa;
-  border-radius: 10px;
-  padding: 5px;
-
-  input,
-  select {
-    border: none;
-    font-size: 16pt;
-  }
-
-  & > .group > label {
-    display: block;
-    min-height: 2em;
-    margin-top: 0.5em;
-    opacity: 80%;
-  }
-}
-
-button:hover {
-  transform: translate(0, -1px);
-  transition: transform 0.1s;
-}
-
-header.top {
-  background-color: #ffffff;
-  box-shadow: 0px 2px 14px 2px rgba(0, 0, 0, 0.1);
-  height: 88px;
-  padding-left: 56px;
-  padding-right: 56px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  h1 {
-    margin: 0;
-    margin-right: auto;
-
-    .title {
-      color: #ea860b;
-      font-weight: bold;
-      font-size: 36px;
-      line-height: 42px;
-      font-family: "Roboto";
-    }
-
-    & > .logo {
-      transition: transform 0.4s;
-      display: inline-block;
-      margin-right: 0.4em;
-      &:hover {
-        transform: rotate(-10deg) translate(0, 8px);
-      }
-    }
-  }
-
-  ul.nav {
-    font-family: $font-family;
-    list-style-type: none;
-    padding: 0;
-    margin: 0;
-    & > li {
-      display: inline;
-      margin-right: 40px;
-      &:last-child {
-        margin-right: 0;
-      }
-      a {
-        color: $text;
-        opacity: 0.6;
-        text-decoration: none;
-        font-style: normal;
-        font-weight: $normal;
-        font-size: 20px;
-
-        &:hover {
-          opacity: 1;
-        }
-
-        &.active {
-          opacity: 1;
-          color: $text-darker;
-          font-weight: $bold;
-        }
-      }
-    }
-  }
-  .walletStatus {
-    margin-left: auto;
-
-    .connection {
-      // padding: 8px;
-      // border: 1px solid;
-      // border-radius: 10px;
-      &.success .indicator {
-        filter: invert(1);
-      }
-    }
-    img.logo {
-      height: 1.5em;
-    }
-  }
-}
-
-main {
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: flex-start;
-}
-
-.swapArea {
-  border-radius: 10px;
-  max-width: 420px;
-  width: 100%;
-  margin-top: 100px;
-
-  border: 1px solid #999;
-  padding: 1em;
-
-  nav {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-    width: 100%;
-    & > a {
-      font-size: calc(16px + 2vmin);
-      opacity: 50%;
-
-      &.active {
-        opacity: 100%;
-      }
-    }
-  }
-
-  & > hr {
-    margin-bottom: 1em;
-  }
-
-  & > form {
-    display: flex;
-    flex-direction: column;
-    align-content: stretch;
-
-    & > * {
-      margin-top: 1em;
-    }
-
-    & > button:last-child {
-      min-height: 2.5em;
-    }
-  }
-}
-
-form.earn {
-  & > span {
-    text-align: center;
-    margin-bottom: 0.3em;
-
-    &.apy {
-      font-size: 3em;
-      display: block;
-      margin-bottom: 0.3em;
-    }
-  }
-}
