--- src/history.ts
@@ -1,2 +1,0 @@
-import { createBrowserHistory } from "history"
-export default createBrowserHistory()
